bugfix: use bigger trace space buffer, to handle large contog vs contig/read aligments

+++ lib/pass.h
@@ -41,8 +41,8 @@ typedef struct
     off_t off_start;
     off_t off_end;
 
-    int tmax;
-    int tcur;
+    uint64 tmax;
+    uint64 tcur;
     ovl_trace* trace;
 
     size_t tbytes;
